Move expose helpers to separate module to fix test import
The Clipboard import in postUtils.tsx broke postUtils.test.ts in CI
because react-native modules can't load in the Vitest environment.
Moved getExposeReferencePath and useHandleExposeSuccess to exposeUtils.ts.
