+++ packages/app/ui/utils/index.ts
@@ -2,6 +2,7 @@ export * from './channelUtils';
 export * from './clipboardUtils';
 export * from './groupUtils';
 export * from './colorUtils';
+export * from './exposeUtils';
 export * from './postUtils';
 export * from './user';
 export * from './animation';
