+++ packages/app/ui/utils/postUtils.tsx
@@ -1,7 +1,5 @@
-import Clipboard from '@react-native-clipboard/clipboard';
 import * as db from '@tloncorp/shared/db';
-import { useToast } from '@tloncorp/ui';
-import { useCallback, useMemo } from 'react';
+import { useMemo } from 'react';
 
 interface ReactionListItemUser {
   id: string;
@@ -85,25 +83,6 @@ export function useReactionDetails(
   return useMemo(
     () => computeReactionDetails(reactions, ourId),
     [reactions, ourId]
-  );
-}
-
-export function getExposeReferencePath(post: db.Post) {
-  const [kind, host, channelName] = post.channelId.split('/');
-  const postId = post.id.replaceAll('.', '');
-  return `/1/chan/${kind}/${host}/${channelName}/msg/${postId}`;
-}
-
-export function useHandleExposeSuccess() {
-  const showToast = useToast();
-  return useCallback(
-    (message: string, url?: string) => {
-      if (url) {
-        Clipboard.setString(url);
-      }
-      showToast({ message, duration: 2000 });
-    },
-    [showToast]
   );
 }
